+++ tests/jwt-renewal.json
@@ -42,20 +42,22 @@
             }
         },
         "views/test/app/renew-jwt": {
-            "view": "fostgres.sql",
+            "view": "odin.app.secure",
             "configuration": {
-                "host": ["request", "headers", "__pgdsn", "host"],
-                "user": ["request", "headers", "__pgdsn", "user"],
-                "dbname": ["request", "headers", "__pgdsn", "dbname"],
-                "sql": [
-                    {
-                        "return": "object",
-                        "path": [],
-                        "GET": {
-                            "command" :"SELECT 1 as result"
+                "secure": {
+                    "view": "odin.jwt.renewal",
+                    "configuration": {
+             
+                    }
+                },
+                "unsecure": {
+                    "view": "fost.response.401",
+                    "configuration": {
+                        "schemes": {
+                            "Bearer": {}
                         }
                     }
-                ]
+                }
             }
         }
     }
